Reapply "ox-html: Use <img> tags for svgs, not <object>"
This reverts commit 302bc6393a11bae823da297c073f76633a591721.

It turns out that <object ...> tags with data pointing outside the
website are not rendered by the modern
qtwebengine (linked svgs in https://orgmode.org/tools.html are not
rendered). Even though Firefox does render the same page, it no longer
appears safe to use <object ...> for svg.  So, we have to use <img...>
with its limitations.