ox: Allow anonymous parent backends
* lisp/ox.el (org-export-get-all-options):
(org-export-get-all-filters): Do not demand parent chain to be from
registered backends.  Allow anonymous parents as well.

This is necessary because `orgtbl-to-generic' now uses anonymous
backend via normal ox.el machinery and ob-gnuplot layers an anonymous
backend on top (which is not invalid, given that `orgtbl-to-generic'
did work with this setup in the past).