org-agenda-write: Remove redundant `set-buffer'
* lisp/org-agenda.el (org-agenda-write): Drop `set-buffer' call at the
end.  It does nothing since the rest of the code is wrapped in
`save-excursion' and the aim of the `set-buffer' is to switch to the
agenda buffer we just saved.  Ignore AGENDA-BUFNAME argument - it is
unused now.  Update the docstring, removing the mention of what
appears to be the now-ignored argument.
(org--batch-store-agenda-views): Update the caller.  Passing current
agenda buffer name is no longer necessary.  We set it before the call
anyway.

Reported-by: Sławomir Grochowski <slawomir.grochowski@gmail.com>
Link: https://orgmode.org/list/87ed7dwygq.fsf@gmail.com