org.el: Add alias for `org-insert-structure-template' using "block" term
* lisp/org.el (org-insert-block-template): Create as an alias
of `org-insert-structure-template'.

Markup structures like "#+begin_src" and "#+being_quote" can be
inserted using the command `org-insert-structure-template'.  In much
of the documentation, these structures are called "blocks", including
the existing documentation for `org-insert-structure-template'.
Therefore, add the alias `org-insert-block-template' as a more
discoverable name in case a user is familiar with the term "block"
but not with the term "structure".

TINYCHANGE