Backport commit a3b308308 from Emacs
* doc/org-manual.org (External Links): Change HTTP example to use HTTP.
Changing it to HTTPS has been done before, but it's a mistake.  To avoid
making this mistake again, change the URL to orgmode.org to make it more
similar to the example on the next line, thereby contrasting the two.

; Don't use HTTPS in HTTP-specific example
a3b3083081299049b2ac2333d078c9cac62f2be4
Stefan Kangas
Mon Feb 3 14:03:43 2025 +0100