org-store-link: Do not query about store function when INTERACTIVE? is nil
* lisp/ol.el (org-link--try-link-store-functions):
Do not use interactive query when INTERACTIVE? is nil.
* etc/ORG-NEWS (~org-store-link~ no longer asks to select store
function when called noninteractively): Announce the breaking change.

The original feature prompting the user was introduced in
https://list.orgmode.org/orgmode/87haqbbpsc.fsf@bernoul.li/
However, always prompting becomes problematic when link is stored for
internal purposes and the user does not expect prompts.

In the past, such situation was not frequent to cause the trouble, but
now, when we moved some internal link types (org-id) to define link
via the normal custom link system, clashes between internal and
contributed link types become more likely. In future, more internal
links will probably use `org-link-set-parameters', so this breaking
change is hopefully an overall improvement.

Reported-by: Hermann Graf von Westerholt <hermann.westerholt@rwth-aachen.de>
Link: https://orgmode.org/list/27BFA4F7-22BB-42A8-97B4-F3590C015BDF@rwth-aachen.de