org-element: Fix parsing top-level property drawer after comment and blanks
* lisp/org-element.el (org-element--current-element): Trust parser
mode when deciding whether we are at property drawer or not.
Requiring blank lines + `bob' is not accurate are we may have comment
lines before top-level drawer.  Also, the other check (preceded by
non-blank line) was not exactly precise and still relied upon parser
mode to be set correctly.  So, let's do it all the way.
* testing/lisp/test-org-element.el (test-org-element/property-drawer-parser):
Add new test case.

Reported-by: @inkbottle:matrix.org