Fall back to Fundamental mode when editing code
* doc/org-manual.org (Editing Source Code): Describe the change in
behavior, namely how Org falls back to Fundamental mode when
appropriate major mode is unavailable, instead of showing an error
message.
* etc/ORG-NEWS: (Source blocks fall back to Fundamental mode, New
function ~org-src-get-lang-mode-if-bound~): Describe the new
functionality and the new function that enables it.
* lisp/ob-tangle.el (org-babel-tangle):
* lisp/org-lint.el (org-lint-suspicious-language-in-src-block):
Simplify code with the new function `org-src-get-lang-mode-if-bound'.
* lisp/org-src.el (org-src-lang-modes, org-src-font-lock-fontify-block)
(org-src-get-lang-mode, org-src-get-lang-mode-if-bound)
(org-edit-export-block, org-edit-src-code, org-edit-inline-src-code):
When editing source blocks, inline source code, or export blocks and
the major mode listed in `org-src-lang-modes' is not available, do not
stop with a user error.  Instead, fall back to the Fundamental mode,
with an informational message.

Done in preparation for the GraphViz mode support:

Reported-by: Rudolf Adamkovič <rudolf@adamkovic.org>
Link: https://list.orgmode.org/m2zfikfhgp.fsf@adamkovic.org/