lisp/org-refile.el: Replace links with their descriptions consistently
* lisp/org-refile.el (org-refile-get-targets): Replace links with
their descriptions in refile targets when
`org-refile-use-outline-path' is nil.  This is to make things
consistent with how the targets are displayed when the variable is
non-nil.

If an org refile target was a link, the refile interface would show
all of: the link, the description and the surrounding brackets. That
was quite difficult to read in the compressed space of the minibuffer.

This change replaces the full link with just the description, which is
easier to read.

A similar change was discussed on the mailing list in 2008, and
apparently accepted -- but somehow lost from the refile code:

Link: https://list.orgmode.org/6CB4F483-0332-4F40-9CDF-99C90A9BD2B1@uva.nl/

TINYCHANGE