lisp/org.el: Allow `org-yank-image-save-method' to be a function
* lisp/org.el (yank-media): Allow `org-yank-image-save-method' to be
a function to be called to produce the directory path.  Also make
sure that there just a single `(insert)'.
(org-yank-image-save-method): Allow function as a value.
* doc/org-manual.org (Drag and Drop & ~yank-media~): Synchronise
documentation with the changes in org.el.
* etc/ORG-NEWS (New and changed options): Document the fact that
`org-yank-image-save-method' can now be a function.

`org-yank-image-save-method' had no way to obtain a directory to save
clipboard contents dynamically.  It could either specify that the
yank location should be determined by the attachment machinery, or
it could be a global fixed directory name.

Current change allows setting up `org-yank-image-save-method' to a
function which would be called to determine where to save an image,
and could take into account buffer's name, whether it is remote or
local, et cetera.