org-fold: Avoid using 'outline overlay spec
* lisp/org-fold.el (org-fold-initialize): Use org-fold-outline as
invisibility spec.  `reveal-mode' treats 'outline special (because of
symbol property set by `outline-mode').  Because `outline-mode'
folding is not compatible with org-fold, we need to use something
else.
* testing/lisp/test-org-list.el (test-org-list/move-item-down):
(test-org-list/insert-item): Update tests to check for the new spec.

Note that it should not be too breaking in practice to change this.
(1) 'outline remains an alias in org-fold-core API calls
(2) We had no reports about changed spec names when switching between
text properties and overlays for folding, so those are not problematic
in practice.

Reported-by: Brett <brett.kerwin1@gmail.com>
Link: https://orgmode.org/list/CACZEyLj83MEyd0NvPTM9=iotsC9h4q7eDTLf+w-1jfy4LvPMiQ@mail.gmail.com