lisp/ox-html.el: Convert numeric to named character references
* lisp/ox-html.el (org-html-special-string-regexps, org-html--tags,
org-html-format-headline-default-function, org-html-timestamp,
org-html-table-cell, org-html-verse-block): Replace numeric character
references with their named equivalent. This should make the code
clearer since the reader only needs to remember character names instead
of hexadecimal Unicode code points. The equivalency table can be
found in section 13.5 of the HTML standard:
https://html.spec.whatwg.org/multipage/named-characters.html#named-character-references.
All used named character references were introduced in HTML 4.0 (or
earlier) which means that they should be supported in XHTML as well.
(org-html-checkbox-types): Change &#xa0; to &nbsp; as
elsewhere. Unfortunately, &#x2610; and &x#2611; don't have a
corresponding named character reference in the HTML standard, but it
should be clear enough what they are supposed to represent from the
context.
* testing/lisp/test-ox-html.el (ox-html/checkbox-ascii): Adjust to
use of &nbsp; in `org-html-checkbox-types'. Since &nbsp; isn't part
of (vanilla) XML, we need to use `libxml-parse-html-region' which
changes the output a little bit: The surrounding html and body elements
are retained (?) as well as the newlines between the li elements.