org.el: Make `org-map-region' properly set point at first heading
* lisp/org.el (org-map-region): Add a call to `goto-char' to go
to the beginning of the match after the first search.

* testing/lisp/test-org.el (test-org/map-region): Add tests
for `org-map-region'.

The current logic of `org-map-region' uses two searches.

In the first, the point is left after the asterisks of the heading.
In the subsequent calls, the point is correctly set to the first
asterisk of the headline.

This change makes sure the first header is processed the same way
as the rest.

TINYCHANGE