ox-html: Fix TOC generation for non-toplevel headlines
* lisp/ox-html.el (org-html--toc-text): Correctly set initial level
based on scope argument and improve docstring.
* lisp/ox-html.el (org-html-toc): Pass scope argument to helper function.
* testing/lisp/test-ox-html.el (org-html/test-toc-text): New tests
for TOC generation.

When exporting a document that starts with a headline of level > 1
(e.g., no top-level headline), the HTML table of contents was
rendered incorrectly.  This was caused by `org-html--toc-text'
calculating the initial nesting level based on the first headline's
level, which is incorrect for a global TOC that should always start
from a base level of 0.

This commit adjusts `org-html--toc-text' to correctly handle this by
checking its optional `scope' argument.