testing: Suppress interactive prompts during testing and on quit
Running the tests interactively would give the user multiple prompts
asking if they wanted to kill a modified buffer.  Then when quitting
Emacs, the user would receive prompts asking if they wanted to save
some of the temporary test files.

* testing/org-test.el (org-test-kill-buffer): New function.
(org-test-at-id, org-test-in-example-file,
org-test-with-temp-text-in-file, org-test-kill-all-examples): Use new
function to kill buffers.
* testing/lisp/test-ob-tangle.el (ob-tangle/tangle-to-self, ob-tangle/bibtex): Use
`org-test-with-temp-text-in-file' to ensure buffers are killed.
* testing/lisp/test-org-agenda.el (test-org-agenda/set-priority): Use
`unwind-protect' for test cleanup and kill a buffer to "agenda-file.org".
* testing/lisp/test-org-capture.el
(test-org-capture/org-capture-expand-olp): Use
`org-test-with-temp-text-in-file' to ensure buffers are killed.

* testing/lisp/test-org-protocol.el
(test-org-protocol/org-protocol-capture-file): Add cleanup for buffers
and files.
* testing/lisp/test-ox.el (test-org-export/org-export-copy-buffer):
Kill buffer.