lisp/ox-html.el: Add missing keywords and `#+OPTIONS:' in backend options-alist
* lisp/ox-html.el (html): Into the HTML backend options-alist, add
missing keywords and `#+OPTIONS:' when possible.  Also for options
with a keyword defined, set the multi-keywords behavior if possible.
* doc/org-manual.org (HTML preamble and postamble, Links in HTML
export, Tables in HTML export, Images in HTML export, Math formatting
in HTML export, CSS support): Only update to manual sections where the
modified options are already described.
* etc/ORG-NEWS (ox-html: More options can be set from an Org-mode
file): List all the new keywords and `#+OPTIONS:' added.  Also list
the options with keyword that accept its keyword being set multiple
times in the same buffer.

In the HTML backend, some options defined in the options-alist have no
keyword or `#+OPTIONS:' defined.  I changed that.  For options with
keyword, I also defined the multi-keyword behavior.  As requested on
the mailing list, I only updated the manual for options that where
already described.  A complete refresh of the manual about all the
HTML backend options is needed, on a separate patch.

Due to a bug, `:html-extension' option cannot be set in buffer.  So,
no keyword added for it.  Also, `:html-divs' are to complex to be
defined by an `#+OPTIONS:' or a keyword.  It will need a modification
on another patch.