Testing: Include other test files use `eval-when-compile'
When using normal `require', all of the tests defined in a file will
be added to the current session.  By using `eval-when-compile', we can
avoid this so tests can be run file by file.

* testing/lisp/test-oc.el: Wrap the require of "test-ox" in
`eval-when-compile'.
* testing/lisp/test-org-habit.el: Wrap the require of
"test-org-agenda" and "test-org" in `eval-when-compile'.
* testing/lisp/test-org-agenda.el (org-test-agenda--agenda-buffers)
(org-test-agenda--kill-all-agendas): Use `defsubst' instead of `defun'
so they can used when required using `eval-when-compile'.