org-fold: Fix regression after Org 9.5.5 when we reveal too much after isearch
* lisp/org-fold-core.el (org-fold-core--isearch-show): Do not be as
aggressive unfolding the isearch overlay.  Unless we encounter the
edge case with point moved outside the invisible overlay, simply use
`org-fold-core-isearch-open-function', which respects
`org-fold-show-context-detail' as needed.  There is no need to reveal
the overlay that is left then.  Note that the bug fixed in 24c4a40f7
is still not reproducible, so we are safe-ish.

Reported-by: dranath D <dranath254@gmail.com>
Link: https://list.orgmode.org/orgmode/CAFeAMKJV86XXeATX13Df8aP0ebbVfjb68wN9Hp6YhKaezr6rkQ@mail.gmail.com/